Allow `cmd` to be a function
This adds support for specifying a function as the `cmd` option of a
linter. This makes it possible to dynamically change the command, for
example based on the presence of a file in a directory.

When the function returns a string, it's executed as-is. When the
function returns a table, the first value is used as the command name;
all other values are added to the start of the arguments list (in
order). This makes it possible to easily switch between (for example)
just `rubocop` and `bundle exec rubocop`.

Fixes #18