Tell golangci to use the current buffer only
If no path is passed to golangci-lint, it will run on the current
folder, yielding many different diagnostics which are not related to the
current file.

The problem is that those are still going to be displayed on the current
buffer, using their position in their files. This lead to a meaningless
superposition of all diagnostics on the current file.

The present changes fixes that by scoping the linter run on the
currently opened file.