golangci-lint: add FromLinter to diagnostic source
golangci-lint is a meta linter composing a number of other popular
linters according the config it reads. It is often helpful to know
exactly which linter produced a diagnostic. In fact, when run on the
command line, the tool prints the linter name by default.

This change the diagnostic source to include the linter. Note that while
golangci-lint accepts `--print-linter-name=false`, this setting appears
to be ignored when `-out-format json` is set.