golangci-lint: appending file breaks some linters
Appending files means that some linters consider only the contents of the given file when linting it. This is a problem as those linters will not be able to find symbols defined in other files in the same package. `typecheck` is the easiest linter to verify as it will claim that types defined on other files couldn't be found.