Ensure column and line numbers remain >= 0
Some of the linters available with golangci-lint return 0 as a column
value if the issue affects the whole line. The golangci-lint integration
in nvim-lint then subtracts 1, which leads to a negative number.

Newer versions of Neovim treat this as invalid. The result is an error
when trying to jump to that line using `vim.diagnostic.goto_next()`.

This commit ensures that any column or line numbers always remain >= 0.

Closes #176