Call `on_done` callback in `on_exit` handler
This works around an issue where the `on_done` callback was not always
called by the `read_output` function.

The most likely cause for this is that some linters do not generate an
EOF (that luv detects), therefore never triggering the `else` branch in
`read_output`.

The `on_done` is now called in two places:

  1. The existing `read_output` branch
  2. The `on_exit` handler

`on_done` has been extended to detect if it has already been called and
do nothing on the second invocation.

Closes #185