rubocop: handle empty files list being returned
The RuboCop linter's JSON output will always have a "files" key with a
list of linted files. nvim-lint looks for the first of these files with
`decoded.files[1]`.

Given that nvim-lint passes `--force-exclusion` to the `rubocop`
binary, it's entirely possible that the "files" value in the JSON data
will be an empty array (`[]`).

With this change, the nvim-lint processing of the RuboCop results will
prevent an error by not attempting to access an entry in the "files"
array that is not there.

A new test has been added to demonstrate the problem and confirm the
fix.

Future ideas:

- make the `--force-exclusion` flag optional
- alert the NeoVim user that the file for the current buffer was
  excluded, lest they believe that the linter processed the file without
  finding any issues