ignore lint results from processes we kill
fixes #388
when we want to run a new lint process, but an existing one is already
running, we kill the existing process. That process then exits
immediately. Until now we took into account the results from that linter
run and reported them to the user. Better to ignore this incomplete run,
to avoid a "blinking effect" for the user, with errors disappearing and
then immediately returning.
This can particularly happen with slow linters (for instance eslint,
which can easily run for one second).