Disambiguate sqlfluff diagnostics
Lint does not provide severity codes in its messages. The original defaulted everything to `ERROR`. This change downgrades regular messages to `WARN` and only flags the code for a parsing error as `ERROR`. This is done because SQLFLUFF formatting will fail if a parsing error occurs, and identifying the error is much easier when a distinction is made between errors and warnings.