Ensure parse functions run in the context of the original buffer
Problem:

People tend to use functions like `vim.fn.expand("%:p")` in PRs which
would be broken because the parse function is called asynchronous after
the linter process finished and in the meantime the user could've
switched buffers.

Solution:

Use `nvim_buf_call` to ensure the parse function is run in the
context of the original buffer.