golangci-lint: clear alt output format paths to overwrite .golangci.yml (#770)
This is a fix for v2 of golangci-lint that clears out any alternative output paths that could be found in the config file. This allows users to set whichever output format they'd prefer in the config file, while ensuring the linter is properly configured for use with this plugin.

In v1, setting the output format would overwrite all other possibly set values in the config file, which is changed now in v2.

Discussion in #761