shellcheck: use file on disk when possible
The motivation for this is to enable shellcheck's `source-path`
directive to resolve the special path `SCRIPTDIR` correctly, see [0],
which only works when passing the input as a filename in the argument
list. To still be able to lint unnamed buffers we specify `stdin = true`
and `-` as the filename as fallback. This works because shellcheck
ignores stdin input when there are filenames in the argument list. This
workaround could be removed when/if shellcheck implements
`--stdin-filename`, see [1].

[0]: https://github.com/koalaman/shellcheck/blob/20d11c1c33df71a375406c981a629269ee5149a2/shellcheck.1.md?plain=1#L278-L284
[1]: https://github.com/koalaman/shellcheck/issues/2735