typos: handle spaces in paths
If you have a path with a space in it, it currently causes an error in
the typos parser because we match only on non-whitespace characters, not
on whole lines.

```
Diagnostics:
1. Parser failed. Error message:
   ...cal/share/nvim/lazy/nvim-lint/lua/lint/linters/typos.lua:29: Expected value but found unexpected end of string at character 49

   Output from linter:
   {"type":"typo","path":"/Users/gib/Library/Mobile Documents/foo.md","line_num":13,"byte_offset":479,"typo":"consitently","corrections":["consistently"]}
```

The typos CLI docs
(https://github.com/crate-ci/typos/tree/master?tab=readme-ov-file#custom)
say that it outputs "jsonlines", which I assume means that splitting on
newline is correct.