checkpatch: favor version of the script supplied by projects
Linux kernel is not the only project that uses checkpatch.pl for patch
validation, other projects (for example qemu) use it too, and their
version is different from the one used by the kernel. Also checkpatch.pl
changes from kernel version to kernel version when working with multiple
versions, so requiring a single version covering everything is not
practical.

Additionally running checkpatch.pl on projects that do not use it and do
not use kernel coding style is counter-productive.

To solve this issue make the linter locate and favor the version of
checkpatch.pl belonging to the project that file being linted belongs
to. Also, if checkpatch.pl is not accessible use "true" as the command
so that there are no errors/notifications and no diagnostic.